requirement.txt
## On mac silicon chips: https://github.com/rusty1s/pytorch_scatter/issues/241

torch==1.10.1+cu102 torchvision==0.11.2+cu102 torchaudio==0.10.1 -f https://download.pytorch.org/whl/torch_stable.html
torch-scatter torch-sparse torch-cluster torch-spline-conv torch-geometric -f https://data.pyg.org/whl/torch-1.10.1+cu102.html
networkx==2.3
pyemd
tqdm
tensorboardX
scipy
pyyaml==5.4.1
easydict
matplotlib==2.2.3
igraph
python-louvain